Move ResponseId source code to material root directory
See previous commit.

+++ src/domain/mesh/element/utils/physical_properties/NDMaterialPhysicalProperties.h
@@ -57,7 +57,7 @@ class NDMaterialPhysicalProperties: public PhysicalProperties<NDMaterial>
     double getCommittedAvgStress(const size_t &iComp) const;
     double getMeanGeneralizedStrainByName(const std::string &) const;
     double getMeanGeneralizedStressByName(const std::string &) const;
-    virtual size_t getComponentIndexFromCode(const std::string &) const;
+    virtual int getComponentIndexFromCode(const std::string &) const;
 
     Matrix getCommittedStrain(void) const;
     Matrix getCommittedStress(void) const;
